+++ testmodul/isilive_test_scenarios_guards.lua
@@ -61,6 +61,7 @@ local REQUIRED_MODULES = {
   "isiLive_killtrack.lua",
   "isiLive_kick_tracker.lua",
   "isiLive_combat_events.lua",
+  "isiLive_pi_tracker.lua",
   "isiLive_death_watch.lua",
   "isiLive_death_alert.lua",
   "isiLive_settings.lua",